refactor!: flock-cli
