+++ flock-cli/Cargo.toml
@@ -20,6 +20,9 @@ clap = "2.33.3"
 ctrlc = "3.1.1"
 futures = "0.3.12"
 lazy_static = "1.4.0"
+log = "0.4.14"
+pretty_env_logger = "0.4.0"
+
 rusoto_core = "0.47.0"
 rusoto_lambda = "0.47.0"
 rusoto_s3 = "0.47.0"
