+++ flock-cli/src/fsql.rs
@@ -15,7 +15,18 @@
 
 use crate::rainbow::rainbow_println;
 use anyhow::{anyhow, Result};
+use clap::{App, AppSettings, ArgMatches, SubCommand};
 use rustyline::Editor;
+
+pub fn command(_: &ArgMatches) -> Result<()> {
+    futures::executor::block_on(fsql())
+}
+
+pub fn command_args() -> App<'static, 'static> {
+    SubCommand::with_name("fsql")
+        .about("The terminal-based front-end to Flock")
+        .setting(AppSettings::DisableVersion)
+}
 
 /// The main entry point for fsql.
 pub async fn fsql() -> Result<()> {
