+++ flock-cli/src/main.rs
@@ -22,6 +22,11 @@ mod s3;
 use anyhow::Result;
 
 pub fn main() -> Result<()> {
+    if std::env::var("RUST_LOG").is_err() {
+        std::env::set_var("RUST_LOG", "info");
+    }
+    pretty_env_logger::init();
+
     #[cfg(feature = "cli")]
     repl::main()
 }
