+++ flock-cli/src/repl.rs
@@ -11,27 +11,19 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-use crate::fsql::fsql;
+use crate::fsql;
 use crate::lambda;
 use crate::nexmark;
 use crate::rainbow::rainbow_println;
 use crate::s3;
-use anyhow::Result;
-use clap::{crate_version, App, AppSettings, Arg, SubCommand};
-use ini::Ini;
-use lazy_static::lazy_static;
-use std::env;
-
-lazy_static! {
-    /// Global settings.
-    pub static ref FLOCK_CONF: Ini = Ini::load_from_str(include_str!("../../flock/src/config.toml")).unwrap();
-    pub static ref FLOCK_S3_BUCKET: String = FLOCK_CONF["flock"]["s3_bucket"].to_string();
-}
+use anyhow::Context as _;
+use anyhow::{anyhow, bail, Result};
+use clap::{crate_version, App, Arg};
 
 #[tokio::main]
 pub async fn main() -> Result<()> {
     // Command line arg parsing and configuration.
-    let matches = App::new("Flock")
+    let app_cli = App::new("Flock")
         .version(crate_version!())
         .about("Command Line Interactive Contoller for Flock")
         .version(crate_version!())
@@ -44,117 +36,27 @@ pub async fn main() -> Result<()> {
                 .help("Sets a custom config file")
                 .takes_value(true),
         )
-        .subcommand(
-            SubCommand::with_name("nexmark")
-                .about("The NEXMark Benchmark Tool")
-                .setting(AppSettings::DisableVersion)
-                .arg(
-                    Arg::with_name("run")
-                        .short("r")
-                        .long("run")
-                        .help("Runs the nexmark lambda function"),
-                ),
-        )
-        .subcommand(
-            SubCommand::with_name("upload")
-                .about("Uploads a function code to AWS S3")
-                .setting(AppSettings::DisableVersion)
-                .arg(
-                    Arg::with_name("code path")
-                        .short("p")
-                        .long("path")
-                        .value_name("FILE")
-                        .help("Sets the path to the function code")
-                        .takes_value(true),
-                )
-                .arg(
-                    Arg::with_name("s3 key")
-                        .short("k")
-                        .long("key")
-                        .value_name("S3_KEY")
-                        .help("Sets the S3 key to upload the function code to")
-                        .takes_value(true),
-                ),
-        )
-        .subcommand(
-            SubCommand::with_name("lambda")
-                .about("The AWS Lambda Tool for Flock")
-                .setting(AppSettings::DisableVersion)
-                .arg(
-                    Arg::with_name("delete function")
-                        .short("d")
-                        .long("delete")
-                        .value_name("FUNCTION_NAME")
-                        .help("Deletes a lambda function")
-                        .takes_value(true),
-                )
-                .arg(
-                    Arg::with_name("delete all functions")
-                        .short("D")
-                        .long("delete-all")
-                        .help("Deletes all lambda functions"),
-                )
-                .arg(
-                    Arg::with_name("list functions")
-                        .short("l")
-                        .long("list")
-                        .help("Lists all lambda functions (all, *, or a single function)"),
-                ),
-        )
-        .subcommand(
-            SubCommand::with_name("fsql")
-                .about("The terminal-based front-end to Flock")
-                .setting(AppSettings::DisableVersion),
-        )
-        .get_matches();
+        .subcommand(nexmark::command_args())
+        .subcommand(s3::command_args())
+        .subcommand(lambda::command_args())
+        .subcommand(fsql::command_args());
 
     rainbow_println(include_str!("./flock"));
 
-    match matches.subcommand() {
-        ("nexmark", Some(nexmark_matches)) => {
-            if nexmark_matches.is_present("run") {
-                nexmark::run()?;
-            }
-        }
-        ("upload", Some(upload_matches)) => {
-            s3::put_function_object(
-                &FLOCK_S3_BUCKET,
-                upload_matches
-                    .value_of("s3 key")
-                    .expect("No function s3 key provided"),
-                upload_matches
-                    .value_of("code path")
-                    .expect("No function code path provided"),
-            )
-            .await?;
-        }
-        ("lambda", Some(lambda_matches)) => {
-            if let Some(function_name) = lambda_matches.value_of("delete function") {
-                lambda::delete_function(function_name).await?;
-            } else if lambda_matches.is_present("list functions") {
-                lambda::list_functions(lambda_matches.value_of("list functions")).await?;
-            } else if lambda_matches.is_present("delete all functions") {
-                lambda::delete_all_functions().await?;
-            }
-        }
-        ("fsql", Some(_)) => {
-            fsql().await?;
-        }
-        _ => {
-            println!("{}", matches.usage());
-        }
+    let global_matches = app_cli.get_matches();
+    let (command, matches) = match global_matches.subcommand() {
+        (command, Some(matches)) => (command, matches),
+        (_, None) => unreachable!(),
+    };
+
+    match command {
+        "nexmark" => nexmark::command(matches),
+        "upload" => s3::command(matches),
+        "lambda" => lambda::command(matches),
+        "fsql" => fsql::command(matches),
+        _ => bail!(matches.usage().to_owned()),
     }
+    .with_context(|| anyhow!("{} command failed", command))?;
 
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_rainbow_print() {
-        let text = include_str!("./flock");
-        rainbow_println(text);
-    }
 }
